requirements.txt
requests>=2.28.0
pytest>=7.2.0
pytest-html>=3.2.0
allure-pytest>=2.12.0
python-dotenv>=0.21.0
pyyaml>=6.0
jsonschema>=4.17.0